Optimize characters resolver to load all films.
