+++ resolver/starwars/film.go
@@ -33,10 +33,28 @@ func (r *filmResolver) Characters(ctx context.Context, f *swapi.Film) ([]*swapi.
 		urls = append(urls, string(url))
 	}
 
-	return getCharacters(ctx, urls)
+	characters, err := getCharacters(ctx, urls)
+	if err != nil {
+		return nil, err
+	}
+
+	if isFieldRequested(ctx, "films") && len(characters) > 0 {
+		urls := make([]string, 0, len(characters) * len(characters[0].FilmURLs))
+		for _, character := range characters {
+			for _, url := range character.FilmURLs {
+				urls = append(urls, string(url))
+			}
+		}
+
+		// Load and Prime cache
+		// NOTE: This is blocking call
+		getFilms(ctx, urls)
+	}
+
+	return characters, err
 }
 
-func getCharacters(ctx context.Context, urls []string) ([]*swapi.Person, error) {
+func getFilms(ctx context.Context, urls []string) ([]*swapi.Film, error) {
 	entry := middlewares.GetLogEntry(ctx)
 	ids := make([]int, 0, len(urls))
 
@@ -51,10 +69,10 @@ func getCharacters(ctx context.Context, urls []string) ([]*swapi.Person, error) 
 		ids = append(ids, id)
 	}
 
-	characters, errs := loaders.GetPersonLoader(ctx).LoadAll(ids)
+	films, errs := loaders.GetFilmLoader(ctx).LoadAll(ids)
 	if len(errs) > 0 && errs[0] != nil {
-		return nil, errs[0]
+		return nil, errors.NewAPIError(errs[0])
 	}
 
-	return characters, nil
+	return films, nil
 }
